🏗️ refactor: version advanced search components and controller under v1
