--- app/components/advanced_search/group.html.erb
@@ -1,35 +1,0 @@
-<fieldset
-  data-advanced-search-target="groupsContainer"
-  data-advanced-search-group-index="<%= @group_index %>"
-  data-advanced-search-legend-template="<%= t("components.advanced_search_component.group", index: "__INDEX__") %>"
-  class="
-    space-y-4 rounded-lg border border-slate-300 px-4 text-sm text-slate-500 dark:border-slate-600
-    dark:text-slate-400
-  "
->
-  <legend><%= t("components.advanced_search_component.group", index: @group_number) %></legend>
-
-  <%= @form.fields_for :groups, @group, child_index: @group_index do |groups_form| %>
-    <% @group.conditions.each_with_index do |condition, condition_index| %>
-      <%= render AdvancedSearch::Condition.new(
-        groups_form: groups_form,
-        group_index: @group_index,
-        condition: condition,
-        condition_index: condition_index,
-        condition_number: condition_index + 1,
-        fields: @fields,
-        operations: @operations,
-      ) %>
-    <% end %>
-  <% end %>
-
-  <div class="my-4 flex justify-end">
-    <%= viral_button(data: { action: "advanced-search#addCondition" }) do %>
-      <%= t("components.advanced_search_component.add_condition_button") %>
-    <% end %>
-
-    <%= viral_button(classes: class_names("ml-2", "hidden": !@show_remove_group_button), data: { action: "advanced-search#removeGroup" }) do %>
-      <%= t("components.advanced_search_component.remove_group_button") %>
-    <% end %>
-  </div>
-</fieldset>
