+++ app/components/advanced_search/v1/condition.html.erb
@@ -1,15 +1,15 @@
 <fieldset
-  data-advanced-search-target="conditionsContainer"
-  data-advanced-search-group-index="<%= @group_index %>"
-  data-advanced-search-condition-index="<%= @condition_index %>"
+  data-advanced-search--v1-target="conditionsContainer"
+  data-advanced-search--v1-group-index="<%= @group_index %>"
+  data-advanced-search--v1-condition-index="<%= @condition_index %>"
+  data-advanced-search--v1-legend-template="<%= t("components.advanced_search_component.v1.condition", index: "__INDEX__") %>"
   data-advanced-search-selected-field="<%= selected_field %>"
-  data-advanced-search-legend-template="<%= t("components.advanced_search_component.condition", index: "__INDEX__") %>"
   class="
     flex items-end space-x-2 rounded-lg border border-slate-300 p-4 text-sm text-slate-500 @max-xl:flex-col
     @max-xl:space-y-2 dark:border-slate-600 dark:text-slate-400
   "
 >
-  <legend><%= t("components.advanced_search_component.condition", index: @condition_number) %></legend>
+  <legend><%= t("components.advanced_search_component.v1.condition", index: @condition_number) %></legend>
 
   <%= @groups_form.fields_for :conditions, @condition, child_index: @condition_index do |conditions_form| %>
     <% invalid_field = @condition.errors.include?(:field) %>
@@ -26,7 +26,7 @@
                                options,
                                { include_blank: true },
                                {
-                                 "data-action": "change->advanced-search#handleFieldChange",
+                                 "data-action": "change->advanced-search--v1#handleFieldChange",
                                  "aria-label": field_label,
                                  "aria-invalid": invalid_field,
                                  "aria-describedby":
@@ -39,7 +39,7 @@
           field: :field,
           options: options,
           data: {
-            action: "change->advanced-search#handleFieldChange",
+            action: "change->advanced-search--v1#handleFieldChange",
           },
           aria: {
             label: field_label,
@@ -67,7 +67,7 @@
                              ),
                              { include_blank: true },
                              {
-                               "data-action": "advanced-search#handleOperatorChange",
+                               "data-action": "advanced-search--v1#handleOperatorChange",
                                "aria-label": operator_label,
                                "aria-invalid": invalid_operator,
                                "aria-describedby":
@@ -80,7 +80,7 @@
       id: conditions_form.field_id(:operator, "error") %>
     </div>
 
-    <%= render AdvancedSearch::Value.new(
+    <%= render AdvancedSearch::V1::Value.new(
       conditions_form: conditions_form,
       group_index: @group_index,
       condition: @condition,
@@ -96,9 +96,9 @@
       text-slate-400 hover:bg-slate-100 dark:bg-slate-800 dark:text-slate-500 dark:hover:bg-slate-700
       dark:hover:text-white
     "
-    aria-label="<%= t("components.advanced_search_component.remove_condition_aria_label") %>"
-    title="<%= t("components.advanced_search_component.remove_condition_aria_label") %>"
-    data-action="advanced-search#removeCondition"
+    aria-label="<%= t("components.advanced_search_component.v1.remove_condition_aria_label") %>"
+    title="<%= t("components.advanced_search_component.v1.remove_condition_aria_label") %>"
+    data-action="advanced-search--v1#removeCondition"
   >
     <%= icon(:x, size: :sm) %>
   </button>
